added a command line tool with exec functionality
This first CLI version implements a simple exec implementation that
reuses code from my rejected conan PR [1] with similar functionality.
However, since the internal conan API is not available another
information source had to be used. Currently the json generator is
employed. Unfortunately, this generator misses variables from the
profile so another way has to be investigated in the future.

[1] https://github.com/conan-io/conan/pull/2516