removed special defaults for profiles and build arguments
Previously profiles and build arguments where optional and special
magic tried to infer values from the command line. However, this
behaviour is not needed most of the time and actively prevents
certain needed invocations. (e.g., --build without argument)

This refactoring removed the magic and delegates the responsibility
to the caller of the API. The default "outdated" build flag is still
used when no argument is specified but can easily be replaced with
any other list.